File restructuring Grouped all related files into their own folders
