+++ src/App.js
@@ -1,11 +1,11 @@
 import React, { useState, useEffect } from 'react'
 import { Route, Switch } from 'react-router-dom'
-import Dashboard from 'pages/Dashboard'
-import MonitorDetails from 'pages/MonitorDetail'
-import Monitors from 'pages/MonitorsPage/MonitorsPage'
+import Dashboard from 'pages/Dashboard/Dashboard'
+import MonitorDetails from 'pages/MonitorDetails/MonitorDetail'
+import Monitors from 'pages/MonitorsPage/Monitors/MonitorsPage'
 import NavBar from 'components/Navbar/Navbar.js'
 import SignIn from 'components/SignIn/SignIn'
-import MonitorsPage from 'pages/MonitorsPage/MonitorsPage'
+import MonitorsPage from 'pages/MonitorsPage/Monitors/MonitorsPage'
 import { useSelector, useDispatch } from 'react-redux'
 import 'App.style.scss'
 import Settings from 'pages/Settings/Settings'
