+++ src/pages/Dashboard/MonitorDetail.css
