+++ src/pages/MonitorDetails/MonitorDetailHeader.scss
