+++ src/pages/MonitorDetails/MonitorDetails.style.scss
