+++ src/pages/MonitorDetails/MonitorsLineChart.styles.scss
