+++ src/pages/MonitorsPage/CalendarFilter/CalFilter.scss
