+++ src/pages/MonitorsPage/Monitors/MonitorsPage.scss
